mono: Convert string code to use custom marshalers
The called from delegates, the string ownership works kinda reverse of
the normal behavior. For example, when we return/out a string to C, by
default the managed code passes the ownership to the unmanaged code.

Also tried to simplify the manual marshalling code and updated test to
reflect new ownership behavior.