evas: Fix use of invalid EO object
During a window destruction, the input devices are deleted and
removed from the internal devices list. Since 83d18617b4375
basic devices are children of a seat, which means that the seat
deletion was triggering the devices deletion as well. As a
consequence, we were walking an invalid list of devices.

I'm not sure we should have a list of children in the device
data and I'm also not sure we should have a list of devices in
evas, instead of just the list of seats. It's a bit messy right
now.

Ref T5540