eolian_mono: Generate the right dllimport argument
Running the C# bindings with VS runtime requires passing the right
libraries names to DllImport, as it does not have something like dllmap
to map between the unix-like name (e.g. 'eina') and the actual filename
in Windows (e.g. libeina-1).

In order to get the version number, we calculate it from the major and
minor version numbers, like libtool does.

This commit changes just the generated code. The handwritten C# code
will be fixed in the next commits.