efl_mono: Refactor the DllImport library naming
Instead of harcoding the names of the libraries all over the code, put
them in a separate class and just reference it from the DllImport call.

This will help dealing with Windows builds, as the string passed to
DllImport is different from inside Unix. (For more info, see the
previous commit on Eolian_Mono changes).