eolian_mono: Add cast and Equals to generated code
When a method returns an EoId we create just an efl.ObjectConcrete and
return it. This behavior could be problematic by requiring to manually
instantiate the desired class through the overload receiving the
underlying IntPtr.

This commits adds a method called static_cast(efl.Object) to ease this
conversion while hiding this implementation detail. Also, an override of
Object.Equals(Object obj) was added and two objects are equal if their
handles are the same.