eolian_mono: Change to PascalCasing for methods
The managed methods now have PascalCasing by default. Whenever a verb is
found to be the last name in a method name, it is reordered to the front
(e.g. text_set becomes SetText). This will be more idiomatic to C#
developers as it's the recommended convention.

This commit also adds a "name_helpers.hh" with some helper functions
to ease this method name handling and may be moved to eolian_cxx later
when the JS bindings start using it.