efl_mono: Generate file with eo library mapping
Instead of passing a huge number of command line arguments, eolian_mono
will parse the mapping between .eo file and library from a csv file in
the format:

filename.eo,library

For efl core libraries, the names will be replaced with the correct
library name from the generated enum efl.Libs. For third parties, use
the actual name to be passed to DllImport.